Crawler_Extractor/requirements.txt
streamlit>=1.30
pypdf>=3.17
pymupdf>=1.24
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
openai>=1.6
tenacity>=8.2
langchain>=0.2
langchain-core>=0.2
langchain-openai>=0.1
langchain-google-genai>=2.0
supabase>=2.3
websockets>=12
duckduckgo_search>=6.1
structlog>=24.1
pydantic>=2.7
tiktoken>=0.7
storage3
